refactor: close state-write evidence ownership
